+++ CHANGELOG.md
@@ -1,4 +1,14 @@
-# Reflectis-PLG-TasksReflectis
+# Release notes
+
+## v2.0.1
+### Fixed
+- Now reverting tasks only reverts them one time. Fixed in `TaskSystemReflectis`
+
+## v2.0.0
+
+### Changed
+
+- Changed package name, from Reflectis-PLG-TasksReflectis to Reflecits-CreatorKit-Tasks, and updated namespaces according to new package name.
 
 ## v1.2.0